Fix issues with progress bar animation
Made it more robust by basically implementing it myself. The framerate will be lower, but that's fine.
The previous implementation had lots of trouble with changing the animation after seeking, pausing, stopping and resuming, continuing to the next track, moving to a different page and then going back, etc... This is simpler and more fault-tolerant.