Background cache of directory contents
This is not enough of a cache to actually work for us. We also need to cache the subdirectories in each directory, and hashes of the file contents. But we also just installed FS-cache on the machine that might do the same for us as long as we visit all directories, so maybe this works well enough. Let's see.